chore(web): Polish LaunchDarky async integration
1. Show the loader immediately as part of custom Skeleton CSS.
2. Prevent any flickering until auth and feature flags are resolved.
