+++ apps/web/src/App.tsx
@@ -1,10 +1,13 @@
 import { AppRoutes } from './AppRoutes';
 import Providers from './Providers';
+import { Gate } from './Gate';
 
 export default function App() {
   return (
     <Providers>
-      <AppRoutes />
+      <Gate>
+        <AppRoutes />
+      </Gate>
     </Providers>
   );
 }
