--- apps/web/src/components/launch-darkly/index.ts
@@ -1,1 +1,0 @@
-export * from './LaunchDarklyProvider';
