+++ libs/shared-web/src/components/index.ts
@@ -1,2 +1,1 @@
 export * from './HubspotForm';
-export * from './SignedIn';
